+++ pyproject.toml
@@ -21,13 +21,14 @@ classifiers = [
     "Framework :: Pytest",
 ]
 dependencies = [
-    'tomli; python_version<"3.11"'
+    'tomli; python_version<"3.11"',
+    "pytest-check",
+    "pytest>=7.2",
 ]
 dynamic = ["version"]
 
 [project.optional-dependencies]
 dev = [
-    "pytest>=7.2",
     "flake8",
     "black"
 ]
@@ -48,4 +49,4 @@ addopts = [
 ]
 
 [project.entry-points.pytest11]
-py3arch = "py3arch.pytest_py3arch"
+py3arch = "py3arch.pytest.plugin"
