+++ tests/conftest.py
@@ -2,8 +2,6 @@ from textwrap import dedent
 
 import pytest
 
-pytest_plugins = "myapp.testsupport.myplugin"
-
 @pytest.fixture
 def create_testset(tmp_path):
     def _create_testset(*module_contents):
